Add files via upload
The dataset contains information about credit applicants. Banks, globally, use
this kind of dataset and type of informative data to create models to help in
deciding on who to accept/refuse for a loan.
After all the exploratory data analysis, cleansing and dealing with all the
anomalies we might (will) find along the way, the patterns of a good/bad
applicant will be exposed to be learned by machine learning models.
We’re dealing with a supervised binary classification problem. The goal is to
train the best machine learning model to maximize the predictive capability of
deeply understanding the past customer’s profile and minimizing the risk of
future loan defaults.